requirements.txt
numpy
pandas
scikit-learn
flask
python-dotenv
mysql-connector-python
ipykernel
pymysql
sqlalchemy
seaborn
matplotlib
plotly
-e .

